+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## 0.3.0 - 2017-09-08
 ### Changed
 - Breaking: document type is now set to `element` for all elements. To access section handles – for entries – or handles – for globals, use ``_source.type.keyword` instead of `_type`.